make round u64 instead of usize

+++ crates/aptos-batch-encryption/src/schemes/fptx.rs
@@ -73,7 +73,7 @@ impl BatchThresholdEncryption for FPTX {
 
     type Ciphertext = Ciphertext<FreeRootId>;
 
-    type Round = usize;
+    type Round = u64;
 
     type Digest = Digest;
